feat(settings): move config action into pane title
Remove the standalone settings header and render the settings.json edit action in the shared pane title row.

Preserve the existing settings bridge behavior and accessible error message from the former header action.

Refs VIM-168
